Version 1.0.5
1. Fix double payments

+++ src/main/java/org/boosty/entity/Subscriber.java
@@ -15,7 +15,7 @@ public class Subscriber {
     private String avatarUrl;
     private long onTime;
     private String status;
-    private int payments;
+    private double payments;
     private int price;
     private boolean subscribed;
